优雅地使用MyBatis XML配置版

+++ chapter-base-plugins/chapter-aggregation/src/main/java/com/example/qixin/service/UsersService.java
@@ -29,7 +29,7 @@ public class UsersService {
         return usersApi.findAll();
     }
 
-    public Long insertByBook(Users book) {
+    public Integer insertByBook(Users book) {
 
         return usersApi.save(book);
     }
